session

+++ PHP/09.Session/CRUD/index.php
@@ -1,16 +1,32 @@
 <?php
+session_start();
 require_once "inc/functions.php";
 $info  = '';
 $task  = $_GET['task'] ?? 'report';
 $error = $_GET['error'] ?? '0';
+
+if('edit'==$task){
+    if(!hasPrivilege()){
+	    header( 'location: index.php?task=report' );
+    }
+}
+
 if('delete'==$task){
+    if(!isAdmin()){
+	    header( 'location:index.php?task=report' );
+	    return;
+    }
 	$id = filter_input( INPUT_GET, 'id', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
 	if($id>0){
 	    deleteStudent($id);
-		header( 'location: index.php?task=report' );
+		header( 'location:index.php?task=report' );
     }
 }
 if ( 'seed' == $task ) {
+	if(!isAdmin()){
+		header( 'location:index.php?task=report' );
+		return;
+	}
     seed();
     $info = "Seeding is complete";
 }
@@ -20,7 +36,7 @@ $roll  = '';
 if ( isset( $_POST['submit'] ) ) {
     $fname = filter_input( INPUT_POST, 'fname', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
     $lname = filter_input( INPUT_POST, 'lname', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
-    $roll  = filter_input( INPUT_POST, 'roll',FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+    $roll  = filter_input( INPUT_POST, 'roll', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
     $id = filter_input( INPUT_POST, 'id', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
     if($id){
         //update the existing student
@@ -37,7 +53,7 @@ if ( isset( $_POST['submit'] ) ) {
         if ( $fname != '' && $lname != '' && $roll != '' ) {
             $result = addStudent( $fname, $lname, $roll );
             if ( $result ) {
-                header( 'location: index.php?task=report' );
+                header( 'location:index.php?task=report' );
             } else {
                 $error = 1;
             }
@@ -68,7 +84,7 @@ if ( isset( $_POST['submit'] ) ) {
             <h2>Project 2 - CRUD</h2>
             <p>A sample project to perform CRUD operations using plain files and PHP</p>
             <?php include_once( 'inc/templates/nav.php' ); ?>
-            <hr/>
+
             <?php
             if ( $info != '' ) {
                 echo "<p>{$info}</p>";
@@ -132,6 +148,6 @@ if ( isset( $_POST['submit'] ) ) {
     endif;
     ?>
 </div>
-</body>
 <script type="text/javascript" src="assets/js/script.js"></script>
+</body>
 </html>
